# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skyrx E-Commerce",
        "description": "API Documentation untuk Skyrx E-Commerce Platform",
        "contact": {
            "email": "skyrxstore@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/auth/google/callback": {
            "get": {
                "tags": [
                    "Social Auth"
                ],
                "summary": "Handle Google OAuth callback",
                "description": "Callback endpoint yang dipanggil oleh Google setelah user authorize",
                "operationId": "googleAuthCallback",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Authorization code dari Google",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to frontend with token"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}